requirements.txt
MySQL-python
Django
django-reversion
Pygments==1.3.1
pyOpenSSL
BeautifulSoap
django-compressor
python-memcached
